# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Roadcapture API
  description: Roadcapture API Description
  version: 0.1.0
servers:
- url: http://localhost:8080
...
paths:
  /albums:
    get:
      tags:
      - albums
      summary: 앨범조회
      description: 앨범조회
      operationId: 앨범조회 - 성공
      parameters:
      - name: page
        in: query
        description: 조회할 페이지입니다. 0부터 시작합니다.
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: 한 페이지에 보여줄 사이즈 수입니다.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: 정렬 기준입니다.
        required: false
        schema:
          type: string
      - name: dateTimeFrom
        in: query
        description: 조회할 앨범의 최소 시각입니다.
        required: false
        schema:
          type: string
      - name: dateTimeTo
        in: query
        description: 조회할 앨범의 최대 시각입니다.
        required: false
        schema:
          type: string
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MCwiZXhwIjoxNjQxMTQ2NTYwfQ.YdOln0Ey03yoKgrtyT8BnAy6AsDdWJZCiIz7hBiTdUc
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/albums1485267483'
              examples:
                앨범조회 - 성공:
                  value: "{\"content\":[{\"id\":6,\"createdAt\":\"2022-01-03T02:01:27.484625\"\
                    ,\"lastModifiedAt\":\"2022-01-03T02:01:27.484625\",\"title\":\"\
                    볼거리가 가득한 국내 여행지 1\",\"description\":\"전국의 아름다운 10대 가로수길 중 하나로\
                    \ 선정된 곡교천변 은행나무길은 현충가 입구에 있습니다.\",\"thumbnailUrl\":\"https://www.test.com/test\"\
                    ,\"user\":{\"id\":1,\"username\":\"user1\",\"profileImageUrl\"\
                    :\"https://www.test.com/test\"},\"viewCount\":0,\"likeCount\"\
                    :1,\"commentCount\":10},{\"id\":22,\"createdAt\":\"2022-01-03T02:01:27.748555\"\
                    ,\"lastModifiedAt\":\"2022-01-03T02:01:27.748555\",\"title\":\"\
                    볼거리가 가득한 국내 여행지 2\",\"description\":\"전국의 아름다운 10대 가로수길 중 하나로\
                    \ 선정된 곡교천변 은행나무길은 현충가 입구에 있습니다.\",\"thumbnailUrl\":\"https://www.test.com/test\"\
                    ,\"user\":{\"id\":2,\"username\":\"user2\",\"profileImageUrl\"\
                    :\"https://www.test.com/test\"},\"viewCount\":0,\"likeCount\"\
                    :1,\"commentCount\":10},{\"id\":38,\"createdAt\":\"2022-01-03T02:01:28.02048\"\
                    ,\"lastModifiedAt\":\"2022-01-03T02:01:28.02048\",\"title\":\"\
                    볼거리가 가득한 국내 여행지 3\",\"description\":\"전국의 아름다운 10대 가로수길 중 하나로\
                    \ 선정된 곡교천변 은행나무길은 현충가 입구에 있습니다.\",\"thumbnailUrl\":\"https://www.test.com/test\"\
                    ,\"user\":{\"id\":3,\"username\":\"user3\",\"profileImageUrl\"\
                    :\"https://www.test.com/test\"},\"viewCount\":0,\"likeCount\"\
                    :1,\"commentCount\":10},{\"id\":54,\"createdAt\":\"2022-01-03T02:01:28.222205\"\
                    ,\"lastModifiedAt\":\"2022-01-03T02:01:28.222205\",\"title\":\"\
                    볼거리가 가득한 국내 여행지 4\",\"description\":\"전국의 아름다운 10대 가로수길 중 하나로\
                    \ 선정된 곡교천변 은행나무길은 현충가 입구에 있습니다.\",\"thumbnailUrl\":\"https://www.test.com/test\"\
                    ,\"user\":{\"id\":4,\"username\":\"user4\",\"profileImageUrl\"\
                    :\"https://www.test.com/test\"},\"viewCount\":0,\"likeCount\"\
                    :1,\"commentCount\":10},{\"id\":70,\"createdAt\":\"2022-01-03T02:01:28.417312\"\
                    ,\"lastModifiedAt\":\"2022-01-03T02:01:28.417312\",\"title\":\"\
                    볼거리가 가득한 국내 여행지 5\",\"description\":\"전국의 아름다운 10대 가로수길 중 하나로\
                    \ 선정된 곡교천변 은행나무길은 현충가 입구에 있습니다.\",\"thumbnailUrl\":\"https://www.test.com/test\"\
                    ,\"user\":{\"id\":5,\"username\":\"user5\",\"profileImageUrl\"\
                    :\"https://www.test.com/test\"},\"viewCount\":0,\"likeCount\"\
                    :1,\"commentCount\":10}],\"pageable\":{\"sort\":{\"sorted\":false,\"\
                    unsorted\":true,\"empty\":true},\"pageNumber\":0,\"pageSize\"\
                    :10,\"offset\":0,\"paged\":true,\"unpaged\":false},\"last\":true,\"\
                    totalElements\":5,\"totalPages\":1,\"number\":0,\"sort\":{\"sorted\"\
                    :false,\"unsorted\":true,\"empty\":true},\"first\":true,\"numberOfElements\"\
                    :5,\"size\":10,\"empty\":false}"
    post:
      tags:
      - albums
      summary: 앨범등록
      description: 앨범등록
      operationId: 앨범등록 - 성공
      parameters:
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MCwiZXhwIjoxNjQxMTQ2NTYwfQ.YdOln0Ey03yoKgrtyT8BnAy6AsDdWJZCiIz7hBiTdUc
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/albums-1363830007'
            examples:
              앨범등록 - 성공:
                value: "{\"title\":\"볼거리가 가득한 국내 여행지\",\"description\":\"전국의 아름다운\
                  \ 10대 가로수길 중 하나로 선정된 곡교천변 은행나무길은 현충가 입구에 있습니다.\",\"thumbnailUrl\"\
                  :\"https://www.test.com/test\",\"pictures\":[{\"createdAt\":\"2022-01-03T02:02:42.428283800\"\
                  ,\"lastModifiedAt\":\"2022-01-03T02:02:42.428283800\",\"imageUrl\"\
                  :\"https://www.test.com/test\",\"description\":\"저번에 이어 이번에도 그 목\
                  적지로 향했습니다.\",\"place\":{\"name\":\"곡교천 은행나무길\",\"latitude\":36.1112512,\"\
                  longitude\":27.1146346,\"address\":{\"addressName\":\"충남 아산시 염치읍\
                  \ 백암리 502-3\",\"roadAddressName\":null,\"region1DepthName\":\"충남\
                  \",\"region2DepthName\":\"아산시\",\"region3DepthName\":\"염치읍\",\"\
                  zoneNo\":\"336-813\"}}}]}"
      responses:
        "201":
          description: "201"
  /albums/{id}:
    get:
      tags:
      - albums
      summary: 앨범단건조회
      description: 앨범단건조회
      operationId: 앨범단건조회 - 성공
      parameters:
      - name: id
        in: path
        description: 조회할 앨범 아이디입니다.
        required: true
        schema:
          type: string
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MCwiZXhwIjoxNjQxMTQ2NTYwfQ.YdOln0Ey03yoKgrtyT8BnAy6AsDdWJZCiIz7hBiTdUc
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/albums-id1265395297'
              examples:
                앨범단건조회 - 성공:
                  value: "{\"id\":6,\"createdAt\":\"2022-01-03T02:01:27.484625\",\"\
                    lastModifiedAt\":\"2022-01-03T02:01:27.484625\",\"title\":\"볼거\
                    리가 가득한 국내 여행지 1\",\"description\":\"전국의 아름다운 10대 가로수길 중 하나로 선정\
                    된 곡교천변 은행나무길은 현충가 입구에 있습니다.\",\"thumbnailUrl\":\"https://www.test.com/test\"\
                    ,\"user\":{\"id\":1,\"username\":\"user1\",\"profileImageUrl\"\
                    :\"https://www.test.com/test\"},\"viewCount\":0,\"likeCount\"\
                    :1,\"commentCount\":10,\"pictures\":[{\"id\":7,\"createdAt\":\"\
                    2022-01-03T02:01:27.494627\",\"lastModifiedAt\":\"2022-01-03T02:01:27.494627\"\
                    ,\"imageUrl\":\"https://www.test.com/test\",\"description\":\"\
                    저번에 이어 이번에도 그 목적지로 향했습니다.\",\"place\":{\"id\":8,\"name\":\"곡교천\
                    \ 은행나무길\",\"latitude\":36.1112512,\"longitude\":27.1146346,\"\
                    address\":{\"addressName\":\"충남 아산시 염치읍 백암리 502-3\",\"roadAddressName\"\
                    :null,\"region1DepthName\":\"충남\",\"region2DepthName\":\"아산시\"\
                    ,\"region3DepthName\":\"염치읍\",\"zoneNo\":\"336-813\"}}},{\"id\"\
                    :9,\"createdAt\":\"2022-01-03T02:01:27.511627\",\"lastModifiedAt\"\
                    :\"2022-01-03T02:01:27.511627\",\"imageUrl\":\"https://www.test.com/test\"\
                    ,\"description\":\"저번에 이어 이번에도 그 목적지로 향했습니다.\",\"place\":{\"id\"\
                    :10,\"name\":\"곡교천 은행나무길\",\"latitude\":36.1112512,\"longitude\"\
                    :27.1146346,\"address\":{\"addressName\":\"충남 아산시 염치읍 백암리 502-3\"\
                    ,\"roadAddressName\":null,\"region1DepthName\":\"충남\",\"region2DepthName\"\
                    :\"아산시\",\"region3DepthName\":\"염치읍\",\"zoneNo\":\"336-813\"}}}]}"
    put:
      tags:
      - albums
      summary: 앨범수정
      description: 앨범수정
      operationId: 앨범수정 - 성공
      parameters:
      - name: id
        in: path
        description: 조회할 앨범 아이디입니다.
        required: true
        schema:
          type: string
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MCwiZXhwIjoxNjQxMTQ2NTYwfQ.YdOln0Ey03yoKgrtyT8BnAy6AsDdWJZCiIz7hBiTdUc
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/albums-id-1088817476'
            examples:
              앨범수정 - 성공:
                value: "{\"title\":\"볼거리가 가득한 국내 여행지!!!\",\"description\":\"전국의 아름\
                  다운 10대 가로수길 중 하나로 선정된 곡교천변 은행나무길은 현충가 입구에 있습니다.\",\"thumbnailUrl\"\
                  :\"https://www.test.com/test\",\"pictures\":[{\"id\":23,\"createdAt\"\
                  :\"2022-01-03T02:02:41.834230700\",\"lastModifiedAt\":\"2022-01-03T02:02:41.834230700\"\
                  ,\"imageUrl\":\"https://www.test.com/test\",\"description\":\"저번\
                  에 이어 이번에도 그 목적지로 향했습니다!!!\",\"place\":{\"name\":\"그 은행나무!!!\",\"\
                  latitude\":36.1112512,\"longitude\":27.1146346,\"address\":{\"addressName\"\
                  :\"충남 아산시 염치읍 백암리 502-3\",\"roadAddressName\":\"\",\"region1DepthName\"\
                  :\"충남\",\"region2DepthName\":\"아산시\",\"region3DepthName\":\"염치읍\"\
                  ,\"zoneNo\":\"336-813\"}}}]}"
      responses:
        "200":
          description: "200"
    delete:
      tags:
      - albums
      summary: 앨범삭제
      description: 앨범삭제
      operationId: 앨범삭제 - 성공
      parameters:
      - name: id
        in: path
        description: 조회할 앨범 아이디입니다.
        required: true
        schema:
          type: string
      - name: X-AUTH-TOKEN
        in: header
        description: 로그인 성공 시 발급받은 액세스토큰입니다.
        required: true
        schema:
          type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyIiwicm9sZXMiOlsiUk9MRV9VU0VSIl0sImlhdCI6MTY0MTE0Mjk2MCwiZXhwIjoxNjQxMTQ2NTYwfQ.YdOln0Ey03yoKgrtyT8BnAy6AsDdWJZCiIz7hBiTdUc
      responses:
        "204":
          description: "204"
...
components:
  schemas:
...
    albums-id-1088817476:
      type: object
      properties:
        description:
          type: string
          description: 앨범 설명입니다.
        title:
          type: string
          description: 앨범 제목입니다.
        pictures:
          type: array
          description: 앨범 사진 리스트입니다.
          items:
            type: object
            properties:
              createdAt:
                type: string
                description: 사진 생성 시각입니다.
              lastModifiedAt:
                type: string
                description: 사진 수정 시각입니다.
              imageUrl:
                type: string
                description: 사진 이미지 주소입니다.
              description:
                type: string
                description: 사진 설명입니다.
              place:
                type: object
                properties:
                  address:
                    type: object
                    properties:
                      region2DepthName:
                        type: string
                        description: 시군구명입니다.
                      zoneNo:
                        type: string
                        description: 우편번호입니다.
                      region3DepthName:
                        type: string
                        description: 읍면동명입니다.
                      roadAddressName:
                        type: string
                        description: 도로명주소입니다.
                      region1DepthName:
                        type: string
                        description: 시구명입니다.
                      addressName:
                        type: string
                        description: 지번주소입니다.
                    description: 장소 주소입니다.
                  latitude:
                    type: number
                    description: 장소 위도입니다.
                  name:
                    type: string
                    description: 장소 이름입니다.
                  longitude:
                    type: number
                    description: 장소 경도입니다.
                description: 사진 장소입니다.
              id:
                type: number
                description: 사진 이미지 주소입니다.
        thumbnailUrl:
          type: string
          description: 앨범 썸네일 이미지 주소입니다.
...
    albums-id1265395297:
      type: object
      properties:
        createdAt:
          type: string
          description: 앨범 생성 시각입니다.
        lastModifiedAt:
          type: string
          description: 앨범 수정 시각입니다.
        description:
          type: string
          description: 앨범 설명입니다.
        likeCount:
          type: number
          description: 앨범 좋아요수입니다.
        viewCount:
          type: number
          description: 앨범 조회수입니다.
        id:
          type: number
          description: 앨범 아이디입니다.
        title:
          type: string
          description: 앨범 제목입니다.
        user:
          type: object
          properties:
            id:
              type: number
              description: 사용자 아이디입니다.
            profileImageUrl:
              type: string
              description: 사용자 프로필 사진입니다.
            username:
              type: string
              description: 사용자 이름입니다.
          description: 앨범 등록자입니다.
        pictures:
          type: array
          description: 앨범 사진 리스트입니다.
          items:
            type: object
            properties:
              createdAt:
                type: string
                description: 사진 생성 시각입니다.
              lastModifiedAt:
                type: string
                description: 사진 수정 시각입니다.
              imageUrl:
                type: string
                description: 사진 이미지 주소입니다.
              description:
                type: string
                description: 사진 설명입니다.
              place:
                type: object
                properties:
                  address:
                    type: object
                    properties:
                      region2DepthName:
                        type: string
                        description: 시군구명입니다.
                      zoneNo:
                        type: string
                        description: 우편번호입니다.
                      region3DepthName:
                        type: string
                        description: 읍면동명입니다.
                      roadAddressName:
                        type: string
                        description: 도로명주소입니다.
                      region1DepthName:
                        type: string
                        description: 시구명입니다.
                      addressName:
                        type: string
                        description: 지번주소입니다.
                    description: 장소 주소입니다.
                  latitude:
                    type: number
                    description: 장소 위도입니다.
                  name:
                    type: string
                    description: 장소 이름입니다.
                  id:
                    type: number
                    description: 장소 아이디입니다.
                  longitude:
                    type: number
                    description: 장소 경도입니다.
                description: 사진 장소입니다.
              id:
                type: number
                description: 사진 아이디입니다.
        commentCount:
          type: number
          description: 앨범 댓글수입니다.
        thumbnailUrl:
          type: string
          description: 앨범 썸네일 이미지 주소입니다.
...
    albums1485267483:
      type: object
      properties:
        size:
          type: number
          description: 한 페이지에 보여줄 사이즈 수입니다.
        last:
          type: boolean
          description: 마지막 페이지 여부입니다.
        totalPages:
          type: number
          description: 전체 페이지 수입니다.
        content:
          type: array
          description: 페이지 요소 리스트입니다.
          items:
            type: object
            properties:
              createdAt:
                type: string
                description: 앨범 생성 시각입니다.
              lastModifiedAt:
                type: string
                description: 앨범 수정 시각입니다.
              description:
                type: string
                description: 앨범 설명입니다.
              likeCount:
                type: number
                description: 앨범 좋아요수입니다.
              viewCount:
                type: number
                description: 앨범 조회수입니다.
              id:
                type: number
                description: 앨범 아이디입니다.
              title:
                type: string
                description: 앨범 제목입니다.
              user:
                type: object
                properties:
                  id:
                    type: number
                    description: 사용자 아이디입니다.
                  profileImageUrl:
                    type: string
                    description: 사용자 프로필 사진입니다.
                  username:
                    type: string
                    description: 사용자 이름입니다.
                description: 앨범 등록자입니다.
              commentCount:
                type: number
                description: 앨범 댓글수입니다.
              thumbnailUrl:
                type: string
                description: 앨범 썸네일 이미지 주소입니다.
        first:
          type: boolean
          description: 첫 페이지 여부입니다.
        empty:
          type: boolean
          description: 리스트가 비어 있는지 여부입니다.
        totalElements:
          type: number
          description: 전체 요소 수입니다.
    albums-1363830007:
      type: object
      properties:
        description:
          type: string
          description: 앨범 설명입니다.
        title:
          type: string
          description: 앨범 제목입니다.
        pictures:
          type: array
          description: 앨범 사진 리스트입니다.
          items:
            type: object
            properties:
              createdAt:
                type: string
                description: 사진 생성 시각입니다.
              lastModifiedAt:
                type: string
                description: 사진 수정 시각입니다.
              imageUrl:
                type: string
                description: 사진 이미지 주소입니다.
              description:
                type: string
                description: 사진 설명입니다.
              place:
                type: object
                properties:
                  address:
                    type: object
                    properties:
                      region2DepthName:
                        type: string
                        description: 시군구명입니다.
                      zoneNo:
                        type: string
                        description: 우편번호입니다.
                      region3DepthName:
                        type: string
                        description: 읍면동명입니다.
                      roadAddressName:
                        type: string
                        description: 도로명주소입니다.
                      region1DepthName:
                        type: string
                        description: 시구명입니다.
                      addressName:
                        type: string
                        description: 지번주소입니다.
                    description: 장소 주소입니다.
                  latitude:
                    type: number
                    description: 장소 위도입니다.
                  name:
                    type: string
                    description: 장소 이름입니다.
                  longitude:
                    type: number
                    description: 장소 경도입니다.
                description: 사진 장소입니다.
        thumbnailUrl:
          type: string
          description: 앨범 썸네일 이미지 주소입니다.